feat: Adding getting username in Windows and UEFI
Added functions for getting the user name in Windows and UEFI.
UEFI will always be "efi" since there is no real concept of different users.
In Windows, we can read the user, which seems to give the same name admin or not, so if running as admin, we add (admin) to the returned user name to make it clear which user is actually running the tool.