quick: Special case for M_FALLTHROUGH with clang 3.4.2
Adding a special case for the M_FALLTHROUGH definition in clang 3.4.2 since it was throwing tons of warnings that don't show in newer versions and I couldn't figure out a better way to define the fallthrough to otherwise silence it. For this case, it defines the comment instead of the attribute.