make: Turning on additional warnings for MinGW build
Turning on -Wextra warnings in the MinGW builds to assist with detecting other errors missed in VisualC. This should help improve overall code quality and help prevent us from leaving bugs in the code.

The only warning that is manually disabled is -Wno-format because this shows up all over for printing uint64_t using the PRI?64 macros available in C99.
The only workaround I could find online was to add a definition before stdio include, but that didn't work, so turning off the warning was required.
This should be low-risk as this warning is otherwise enabled for builds in other configurations which will otherwise catch these errors.
The warning only shows in MinGW to help prevent writing code for Windows environments that don't support "long long", which we don't support as it is.