quick: Changing OS_NAME string length to resolve warnings in Linux
A warning was coming due to possible truncation, which would have been ok. This is because the utsname structure does not define lengths for the character arrays, so I used a very large value (512B) to hold the name and resolve the warning.

A future, better version will dynamically allocate the string to an appropriate length in the future.

Signed-off-by: Tyler Erickson <tyler.j.erickson@seagate.com>