feat: Adding a string concatenation function to wrap snprintf
Adding a string concatenation function to wrap how we are using snprintf to perform this operation and keep other parts of our code cleaner.
A duplicated string is used to resolve a warning seen in GCC (but now VS2019) about using the destination string as an input to snprintf. Now a duplicated string is passed in to keep things happy and the warning silent.
The parameters are close to strcat_s from the C11 Annex K functions.

Signed-off-by: Tyler Erickson <tyler.j.erickson@seagate.com>