bug: Adding an additional check to printing data buffer ascii interpretation
Adding an additional check to the function that prints data buffers to the screen that when generating the printable ASCII output, it also runs an is_ASCII check to help prevent a possible crash. It is unlikely to happen, but certain C type functions are crashing when given a non-ASCII character.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>