bug: Fixing a memory leak when reading the unix/linux username
Fixing a bug that causes a memory leak. Since the rawdata buffer is allocated, but was never freed, this was leaking memory. This was only for certain cases, but it likely affects most if not all current builds since the flags around these functions have been around for a while.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>