feat: Changing ansi escape sequences used for colors based on testing
After reading way too much on this and lots of trial and error, I have implemented a new way to set the terminal colors. This was due to behavior I tested years ago in the original version responding to the "intensity" request.
The code now uses 256 color output as default as that seemed to work on all the platforms I could test. The exception to this is AIX, which will use the original colors and the extended bright color codes since the documentation on AIX says these work and does not mention 256 color code support.
For modern systems supporting ANSI escape sequences, it is basically assumed that 256 color output mode is supported (or at least well enough), so this will work as best it can right now.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>