feat: Adding in code to attempt reading the current terminal to find color capabilities
Adding in reading a few environment variables in an attempt to figure out which method(s) of setting colors is required to get the desired color output. This is far from perfect and complete, but should cover the most common cases with what opensea-common can support today.
Most commonly 256 color is supported, but this can take into account when that is not supported (if we can find/determine that terminal).

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>