feat: Making more changes as linux fbcon only allows 8 colors and ignores bright codes
Making more changes to see if I can get bright colors working in fbcon in the background. Right now it only does 8, but it may be possible to get more using the intensity bit, or possibly the color inversion. I think I've been able to have more color in the past using one or both of these methods before.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>