quick: Changing free to safe_Free in byte_Swap_String
Changing a call to free to safe_Free since Mend SAST seems to think this is a double-free, even though it is not.
The safe_Free macro should help since this is checking if the pointer is non-null before freeing and sets it to null once the free has completed so that if it is attempted to be reused in anyway it has already been changed back to NULL instead of the original pointer value.
This may be optimized out, but that is ok.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>