feat: Adding explicit_zeroes function
Adding a function called explicit_zeroes to give the ability to call a function that will make sure memory is zeroed out without memset being optimized out by the compiler.
This does a lot of checking to try and find a method already available in a platform's libc, but it is far from perfect.
Standardized functions memset_explicit (C23) and memset_s (C11 annex k) are preferred when possible before other solutions are used.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>