quick: A few changes based on cppcheck output to limit variable scope, common function name as a variable, etc
Changing the name of the variable used getting the Windows pagesize since a warning from one analyzer seemed to think it was the system() call. This was just to clean up warnings generated from one of many SAST tools used.

Also changed the scope of a variable and the order things were checked in the byte_Swap_String function to help with style and reduce more warnings from cppcheck.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>