bug: Fixing byteswap string sometimes touching null terminator on long enough ATA MN
Fixing the byteswap string function accidentally messing with the NULL terminator.
I saw this happen on a long ATA MN, but only here.
The change is to limit the loop to stringlen - 1 since that will make sure everythign up to the null, but not including the null will be swapped.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>