feat: Only looking up username when UID is not set to root
The username lookup does not need to be done if the UID is set to zero (root), so this reduces some extra code that causes the /etc/passwd file to be opened.
Since a UID of zero will always be root it is unnecessary to open this file to get this information. So this will just return the string "root" when this is detected.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>