feat: Flagging the ability to read a username off by default
Flagging the ability to get the current user's username to off by default.
This was done because a customer reported a security concern as the function that looks up the username in Linux (and other unix-like systems) opens /etc/passwd which the customer did not like.
Rather than delete this functionalty, I have flagged it off by default so that is can be used when enabled in the build process.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>