clean: Cleaning up clang warning about default case handling all enum values
Fixing a couple clang warnings about default case with an enum.
The short version is we do not need or want a default case here so that if a new architecture is added, then a warning can be generated to tell the developer adding it to handle it in these functions as well.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>