bug: Fixing how our function to replace ctime/asctime was formatting the string
Our replacement for ctime/asctime was using %c which causes a warning about y2k (2 digit year) in some locales.
To fix this I made a more explicit format string that should match what ctime and asctime are expected to return.

In order to get the functionality correct I had to use the setlocale function to change to en_us.utf8 to get the output to match what was defined for asctime and ctime.
This locale change is reverted back to what was previously set before the function returns so any other locale use should not be impacted.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>