feat: Initial changes for conversion warnings and new functions
Many changes are the initial changes necessary for the new conversion warning flags.
There are some new functions to help handle conversions to size_t that can be helpful, especially when allocating memory.

There is a new function to handle getting the number of milliseconds since the Unix epoch as well.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>