feat: Simplifying some macros and adding some new ones for other attributes
Simplifying how the M_FALLTHROUGH and M_ATTR_UNUSED macros are set.
Adding M_DEPRECATED to mark functions/variables/etc as deprecated, M_NODISCARD to mark when a result from a function should be used rather than thrown away, and M_NORETURN for use on functions that do not return (call to exit or something like that).
This will make it easier to maintain code and API changes over time as we can mark old functions deprecated to generate warnings as things change in the future.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>