feat: More correct definition of snprintf for older visual studio
Adding a more correct definition of snprintf for older visual studio (2013).
Previously we just called _snprintf, but the return values do not match the standard and on an "out of space" error, the buffer does not get NULL terminated like the standard version.
This implementation should resolve both of those issues if using an older compiler to compile this code.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>