feat: Adding secure file API. Adding more functions to assist reading in data safely.
This is a lot of changes to improve ovarall security in opensea-common.
The secure file API handles most of the capabilities used by our utilities, except for freopen, rename, and remove. These are not used (or rarely used) so are not part of this initial check-in.
One part still needing to be developed is checking that a directory tree is secure before accessing a file. This is still under investigation for how to implement this capability between Windows and POSIX compliant systems.

Also changed to using Microsoft's _s functions when they are available within some functions as well as C11 Annex K's _s functions when they are available.

This also adds in more functions to assist with reading user data more securely and remove uses of atoi and sscanf to use strtol, etc and other parsing that is better defined with less room for potential errors.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>