make: Adding meson detection for best explicit memset type function
Adding multiple checks to see if an explicit memset is available so that it will not be optimized out by the compiler.
While the code has lots of preprocessor checks, this further enhances the check at the build system to make sure it is detected and used when available to make sure things that need to be securely removed from memory are when requested.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>