quick: Moving C++98 nullptr template class outside of extern C
Some old compilers are warning or generating errors because the class was inside the extern C, so moving it to the end of the file, outside the extern C should resolve that issue.

Warnings are temporarily disabled in GCC and Clang for these old versions which are describing a conflict with C++11's nullptr, but we don't need this warning as we are checking the CPP version and if the M_NULLPTR macro was already defined earlier in the same file or not so it will not create a conflict in newer C++ versions.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>