feat: Adding functionality for fprintf with secure file and function to validate format arg
Adding in fprintf functionality for securefile.
Added a file (validate_format.c) to perform the same kind of validation of a format string as the _s MSFT and C11 Annex K requires so it can be used within securefile.

Added a few other attribtues to help optimize the compiler and output warnings when a string literal is not passed to this function so that -Wformat-literal can be generated.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>