feat: Adding detection of when a user has run as sudo to validate directories
Added reading the environemnt variable SUDO_UID to detect when the user has run the application as sudo.
This allows us to trust the user's directories since they were the one requesting the elevation so that their directories can be trusted when validating the full path to the file.
In order to read the SUDO_UID securely, secure_getenv will be used and environ is check for any duplicates of a variable. Any variable with the same name that is duplicated will be considered a tampered environment and the SUDO_UID will not be used to verify directory security.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>