feat: Setting errno to zero before calling some functions
ISO C secure coding documentation says to set errno to zero before calling some functions in the C stdlib so adding those as is recommended for expected results when interpretting the errno for errors after calling them.
Many are strtol type, but there are a few others that are recommended as well.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>