feat: safer versions is is<> and to<> from ctype and _len versions of some funcs
the ctype isspace, isblank, isprint, etc have undefined behavior if passed a value outside of range of unsigned char.
These functions make sure the value is in range before calling the stdlib function and return appropriate values when out of range, even setting errno to ERANGE.

Also updating some of the functions to remove whitespace or find first occurance of one string in another to take a passed-in length value to make them more efficient when the size of the string is already known by the caller.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>