feat: Safe_<>alloc functions and safe memmove and memcpy
Adding safe malloc, calloc, aligned_malloc, etc functions that check if the requested size is zero. If it is zero, a null pointer is returned. In the standards it is undefined what to do on allocation size of zero so these remove that ambiguity and require a non-zero size.
The safer calloc functions check that multiplying size and count will not overflow a size_t during allocation before making a call. If they do, they return a NULLPTR instead.
The safer aligned allocation functions will round all allocations up to the nearest power of 2 with a minimum of sizeof(void*). They also round the allocation size up to the nearest multiple of the alignment to make sure it is compatible with the C11 aligned_alloc requirements and this should be compatible with all similar aligned allocation function requirements.

Last is adding in safe_memmove and safe_memcpy which perform the same checks as the _s versions in C11 annex k and Microsoft's version of these functions.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>