bug: Changing DECLARE_ZERO_INIT_ARRAY
Changing the way a zero initialized array is declared because the original was a GNU extension.
The GNU extension will be used when __GNUC__ is defined, otherwise other methods are used for C++, C99, and a fallback to the old method we've been using if that fails for any reason.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>