feat: Adding asprintf and vasprintf functions
Adding in definitions of asprintf and vasprintf to use when they are not available.
These were originally linux unique, but also show in the dynamic memory technical report for C11.
These can be really useful, so added our own versions to use when not available from the system libc (Windows, possibly other systems too)

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>