feat: optimizing common_string_concat, adding safe_strlen, renaming a couple functions
Optimized the implementation of common_string_concat to be faster. Added using strcat_s and strncat_s when available and other methods already implemented are not available.

Added safe_strlen as an inline function that attempts to use some GCC built-in's to optimize the maximum length when it is possible to detect how much memory was allocated for the string, otherwise limiting it to RSIZE_MAX.

string_n_length renamed to safe_strnlen to make it more clear that this includes additional checks on the string pointer and length provided.

renamed string tokenizing function to safe_string_token to make it more clear that parameters are checked before using them like required by strtok_s in C11 annex K. A macro is included to make transition to the new name easy in this case since it was in use before this change.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>