bug: Fixing some bugs in the way Windows uses the secure file API
Windows has a few unique changes to make this API work as expected.
Certain folders in Windows have different permissions so we needed to add some workarounds for those since that is the way Windows sets itself up out of the box.
These workarounds are added on a per-folder basis of what we know is different from lots of testing and investigation. Not all Windows system folders are handled, only those on the way to the user's documents folder.
Also adding a few null pointer checks when cleaning up secure file since there are cases where certain sub-structures may not be allocated.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>