feat: Adding compile time flag to allow disabling directory security check while migrating
Adding a flag to allow disabling directory security. Doing so will emit a warning message from a pragma.
This is to allow software using opensea-common, but is not yet ready for this feature, to turn it off for now.
This feature will be removed in the future.
Add DISABLE_SECURE_FILE_PATH_CHECK to the preprocessor definitions to disable this for now while migrating to the secure file API.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>