feat: Adding macro/generic/template to make call to safe_Free not require as much casting
Adding a macro/c11 generic/template(C++) to call safe_Free without the caller needing to do the cast to void**.
The new easier macro/template is safe_free (all lowercase), same for aligned and page_aligned versions.
I realized the casting could be worse and more of a mess than a easier generic call like this.
The function does a good job with type enforcement, however the required casts can be more dangerous...rather than going all the way back to the original macro, I think this is a reasonable solution that makes things easier to use and call and it less prone to blindly casting and creating errors.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>