feat: Adding bounds checking versions of bsearch, qsort, lsearch, and lfind
Added in bounds checking versions of bsearch, qsort, lsearch, and lfind.
These are all safe_<func> or safe_<func>_context depending on the version used.

For versions that need a context during comparison there are versions of these that accept a context parameter and for versions that do not need context either null can be passed for the context or a version without a context can be used instead.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>