feat: Adding initial code to set an output error for secure file
Adding initial code to handle setting an error message to provide what went wrong when attempting to access a file and finding incorrect ownership or permissions during evaluation.

This will provide a user more information to determine where the security issue is so that they can address it properly and try again once it has been corrected.

[Seagate/openSeaChest#158]

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>