feat: Making min/max and int round up/down macros faster
Rewrote the min/max and int round up/down macros and functions to be more performant where possible using math operations instead of conditionals when possible.
The int round up/down are more optimized for unsigned values since that same math is undefined in the standards on signed values. This allows for better performance in these small operations which can happen a lot in our code.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>