feat: Removing C11 annex k use for our own functions and fixing clang-tidy warnings
Originally we were relying on C11 annex K when it was available to perform bounds checking, but due to complexity and issues with setting and calling the constraint handler, I decided to remove it in favor of our own implementation.

So this updates a lot of how this works to remove the C11 annex K functions and instead use our own bounds checking. I implemented a few missing functions that were easy to add and switched to them where necessary.

Clang-tidy also found some things to change for more consistent behavior and reducing possible bugs so those changes were also implemetned.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>