feat: Improved performance with print_Data_Buffer functions
Improved performance significantly by creating a line to output one at a time in a single pass and printing it with fputs.
This results in a major performance improvement to this function for big and small buffers. At around 512B the difference is ~16000us (before this change) to ~1500us after this change.
Larger buffers see even more improvement.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>