feat: Adding extra information to safe_ functions when constraint handler is invoked
Came up with a way to pass the function, file, line number, and expression triggering a constraint handler to output this information.
I added a macro to detect when intellisense is active to show an inline function as the hint when using an IDE, but at compile time a macro is used instead to pass this extra information to the real implementation function.

I've added this for all functions that invoke the constraint handler to make it a little easier to backtrack to the source of an issue when something goes wrong.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>