bug: Fixing file extension comparison in secure file
Fixing the file extension check in secure file. The documentation was showing no ".", however it was expecting a dot character.
This update shows the "." in the example and also accepts inputs for comparison that do not include the . when checking for it.
Continuing to use the . for the extension is recommended.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>