feat: Adding timespec_get for pre-C11 and refactoring code to support this
Pre-C11 does not have the struct timespec and timespec_get functions, however they can be written using the various code we already had to calculate the time since unix epoch.

I broke each method into a different section and wrapped it under this call to make it more like writing to C11 standards.

This simplified the functions a lot and it is much easier to read as well.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>