bug: Fixing options that had an optional way to indicate sector size to allow empty units.
A bug was introduced when option parsing and validation was improved where options that didn't previously require units started to require them.
This is not correct as there are plenty of reasons for default behavior for these options.
To fix this I added allowing an empty string for when no option is present.

[Seagate/openSeaChest#173]

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>