feat: Cleaning up many clang-tidy warnings
Cleaning up a lot of clang-tidy warnings. Also updating clang-tidy configuration file for a couple newer features found in newer versions.
Also cleaning up some warnings when this is compiled on it's own without other projects linking it. These warnings are largly due to mismatching the attribute describing if parameters are write-only on a pointer, which gets accessed and should be a read-write parameter instead.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>