feat: Making separate function for get_strerror which will read the error string into a buffer
Moving some logic for getting the error translation to a buffer to a separate function.
Previously this was within the print_errno_to_screen function, but wanted to be able to reuse this differently.

This maintains the wrapping for various ways to get the error string in a threadsafe way.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>