feat: Refactoring code_attributes.h to use modern standard format when possible
Refactoring how attributes are detected and wrapped within macros to use the standardized [[attr]] form when it is available.
This allows keeping consistent style within the macros and may reduce weird errors parsing two different formats (__attribute__(()) and [[]]).
This includes some wrappers to make sure everything is supported and detected properly in GCC, Clang, and MSVC.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>