Bug: Fixing dllexport when no match found. Fixing M_ATTR_UNUSED. Added C++ restrict.
Fixing incorrect definition for when DLL_EXPORT is not defined (was defining DLL_IMPORT).

Fixing M_ATTR_UNUSED for MSVC. The pragmas do not work as expected. Moved these to the M_USE_UNUSED.
Made use of MSVC_PRAGMA wrapper as well.

Added definitions of restrict (__restrict and __restrict__) for C++.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>