bug: Adding workaround for more compilers that don't properly support namespaced C attribtues
__has_c_attribute(ns:attr) is not properly handled by some compilers even though it is allowed in the standard.
MSVC does NOT support this at this time so it's turned off entirely.
Clang may not support it depending on the version. I added a test step in the meson build to detect this. If we can figure out a more specific version, then we can add that to the .h file too.

Also fixed how support for _Generic is detected.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>