feat: Adding clang diagnose_if attribute and implementing it for safe_ functions
Added use of diagnose_if in order to help catch misuse of the safe_ functions for the cases the compiler can detect.
This will generate a compiler error if the conditions that are checked fail.

I also found and fixed a few incorrect checks in some of the less-use functions like safe_lsearch/safe_lfind and updated them as well to be correct and match these diagnose_if conditions.

This attribute cannot diagnose all conditions, especially when there are layers of abstraction, however it will catch some which will improve usage of these functions.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>