feat: Adding nodiscard with a reason message
Adding ability to specify a reason to not discard a return value to code attributes. If a reason cannot be passed due to old attributes or standards, then it is ignored and reverts to M_NODISCARD behavior so that it can still generate at least some kind of warning about not discarding the result.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>