feat: Using strftime_l functions to reduce chances of issues with changing local in multiple threads
Switching to using the strftime_l/_strftime_l when they are available to change the locale instead of the C standard method. The standard method changes the global locale which could possibly change multiple times in a multithreaded environment.
Rather than deal with that, it makes more sense to provide a locale to a function that takes one so that a global modification is not necessary.
This is implemented to POSIX standards and with what is available in Windows.

Some day we should create some kind of locale helper file for this kind of thing, but for now this only exists in this file.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>