feat: Simplifying packed structs for MSVC and adding packed union
Simplifying the packed structures for MSVC to call the packed_aligned version since it's the same thing with an alignment of 1.
Also switching from __pragma to the MSVC_PRAGMA macro which is slightly more compatible between MSVC and clang (although clang shouldn't hit this case).

Added packed unions for aac raid support since it uses it.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>